SCRIPTS: karabo install skips pip dependencies
The `karabo install` command is expected to run in an isolated environment.

The dependencies are expected to be handled using the `DEPENDS` file
or by bundling the packages together manually.

With this MR one could install the `pytest-cov` package on a karabo
installation directly.

The package is available here imported from the external source
repository:
https://git.xfel.eu/gitlab/karaboDevices/pytest-cov

and with this MR, the installation would not require any adjustment
on the repository to be installable on the XFEL CI or in production.