Add test for order between emitted signals and slot calls to same slot
Since the broker message part (and not the inner process shortcut) shall be tested,
the test needs two C++ processes and thus cannot be done with the C++ integration
tests. Python integration tests are the way out, but since one cannot provide C++
device code especially for these (as can be done for Bound and MDL), the test
specific device code is added to the PropertyTest device of the framework.