FIX: Drop too late or too early data on InfluxDb
InfluxDb has an indexing over time and sharding features.
This features can trigger massive delays on the backend, expecially if
the time discrepancies are on the sharding boundary where
influx will be forced to pack and unpack a large amount of data.

This MR fixes this by introducing a limit to the number of
seconds a datapoint will be early or late. Data outside this
boundary will be dropped.