MDL: add __classId attribute to Hash-like elements
The Hash serialization in the C++ api adds a `__classId` attribute
to the hash elements that are derived from hash but are not a hash.

This MR aligns the 2 APIs and enables all bound devices to
process of images and NDArray coming from a MDL device.

This misalignement has stopped the adoption of NDArray in the MDL API
since they did not appear in the DAQ as expected.