BUILD: cmake build does not use the /tmp directory
I found issues in building with cmake, it was constantly
running out of space on my /tmp partition. This cause the build to fail
and I was left with a broken var directory.

This MR addresses this issue by making scoping the
cmake source directory in the `src` folder and allowing the rest to be
the build "environment".