DEPS: Update python and jupyter libraries
ipython is only upgraded to version 7.23.1
version 8 does not play well with ikarabo.

The shell will set a new event loop and it will crash with ikarabo.
At a future date we should figure out how to use ikarabo as an
asyncio loop in ipython.

now, jupyter-notebook, ikarabo and the gui console work