DEPS: python to use tk
The build of python was not creating the `tkinter` module because the CI images did not contain the headers for `tcl` and `tk`.

This is addressed in https://git.xfel.eu/Karabo/ci-containers/-/merge_requests/33

Changes in : `.gitlab-ci.yml`
The current gitlab CI behaviour is to cache an image (to prevent exceeding the pull limit on docker hub).
Consequently, to use a newer CI image we have to use new CI tags. CI images tags are visible here: https://hub.docker.com/r/europeanxfel/karabo-ci/tags

Changes in : `extern/resources`
The version of the `gmock` package, we ship with Karabo relied on the OS's python and uses python2 syntax.
This is not available in many operating systems.

The package `freetype` collides with the expectations of the operating system (`libfontconfig` will point to symbols that we do not ship).
I believe we should not ship this package with Karabo anymore, since we removed the Qt dependency.

The python package `mccabe` is not used in the framework. We should get rid of it.

The python package `setuptools_scm` was not building on my system (green desktop) because the tarball was not a full git repository nor an official tarball taken from PyPI. this is now fixed.

the `Python` package is now compiled with the `--enable-optimizations` option.

[ci-skip]