MDL: Fix file_db project handling
As reported in ticket https://in.xfel.eu/redmine/issues/119691

The file_db is not working as expected.
There seems to be an issue with the namespace definitions.

This MR fixes that. And extends the unit tests to include namespaces.

Many Thanks to Jona Mainberger for having provided a detailed
bug report including the suggestion to the solution here
implemented.