C++: Fix that pipeline re-connection cycle can get broken
Some async handlers that should report success failure or success of a connection attempt
are stored in the InputChannel. If that is destroyed since e.g. GUiServerDevice lost interest
in that connection, the reconnection cycle could get broken and any connection loss (as
triggered by an output channel injecting a new schema) could not be recovered anymore.
Now the cycle is simply restarted regularly without waiting for the unreliable handlers stored
in InputChannel (which require further fixes).
This is more robust and the given the cycle time there is no danger of overlapping cycles
in real life.

Looking into the code of `InputChannel`, I could identify where handlers that should report success or failure of a connection attempt are not called anymore: When the `InputChannel` is destructed meanwhile - and that happens when a GUI client (and thus its server if there are no other clients that have interest) looses interest in an output channel.