GUI: Call updateMatrix on resizing plot.
https://redmine.xfel.eu/issues/132113

We had reverted some changes in pyqtgraph earlier for better performance.
I believe , at that time we did not allow to resize the plot window and
hence left the changes in the changes in 'resizeEvent' as it was.

Now calling 'updateMatrix' method in the 'resizeEvent' which reverts
the changes. All other necessary methods are called in the super class
'resizeEvent' method.

Tested manually and verified resizing the graph both inside and outside the scene
reisizes the data points as well.