Register handlers (especially instanceNew) before topology gathering
Before, there was a time window of about 2 seconds (the time that topology
gathering waits for answers to its slotPing broadcast) where a received
instanceNew is not recognized by the DeviceClient and thus the GuiServer.
And since the underlying SignalSlotable took note of the instance new,
the following heartbeats do not help.