GUI:Crosshair roi at full integer pixel position.
Earlier the position was adjusted by 0.5 (scale/2) and readjusted
while accessing the 'pos' and 'coords'.This was done to place the
crosshair at the middle of pixel square. Now it appears on the
full integer position. The mouse-click position (x,y) is rounded to
get this position.

Also, the crosshair is now translated to only a full integer pixels
on Image.